refactor: shift resolveCategoryLabel to utils folder
