+++ packages/components/src/templates/next/layouts/Article/Article.tsx
@@ -1,12 +1,12 @@
 import { type ArticlePageSchemaType } from "~/types"
 import { getBreadcrumbFromSiteMap } from "~/utils/getBreadcrumbFromSiteMap"
 import { getIndexByPermalink } from "~/utils/getIndexByPermalink"
+import { resolveCategoryLabel } from "~/utils/resolveCategoryLabel"
 
 import { ArticlePageHeader } from "../../components/internal/ArticlePageHeader"
 import { BackToTopLink } from "../../components/internal/BackToTopLink"
 import { renderPageContent } from "../../render"
 import { getTagsFromTagged } from "../Collection/utils/getTagsFromTagged"
-import { resolveCategoryLabel } from "../Collection/utils/resolveCategoryLabel"
 import { Skeleton } from "../Skeleton"
 
 export const ArticleLayout = ({
