+++ packages/components/src/utils/__tests__/resolveCategoryLabel.test.ts
