+++ packages/components/src/utils/resolveCategoryLabel.ts
